Endpoint Working

+++ src/main/java/com/demo/mappingsdemo/onetoone/dto/DriverDto.java
@@ -11,6 +11,5 @@ public class DriverDto {
 
   private String name;
 
-  private int age;
-
+  private Long age;
 }
